Restrict visibility of generated module items
Calculates the appropriate visibility of the structs, methods, and members in the
generated submodule based on the visibility of the structure being modified, so that
they are visibile to the correct level.

In particular, if the base structure's visibility is "relative" (private or `pub([in] super[::...])`),
the submodule's items will have an additional `super` prepended to the visibility path (private becomes
`pub(super)`, `pub(super)` becomes `pub(in super::super)`, etc). "absolute" visibilities like `pub` and
`pub(crate)` are left as-is.

This should fix compiler errors involving the generated submodule "exposing" private interfaces.

Fixes #7